SystemBus::read_u8/write_u8をusize->u16指定に修正

+++ src/nes/eram.rs
@@ -7,13 +7,13 @@ pub struct ExtendedRam {
 }
 
 impl SystemBus for ExtendedRam {
-    fn read_u8(&self, addr: usize) -> u8 {
-        assert!(addr < self.ram.len());
-        self.ram[addr]
+    fn read_u8(&self, addr: u16) -> u8 {
+        assert!((addr as usize) < self.ram.len());
+        self.ram[addr as usize]
     }
-    fn write_u8(&mut self, addr: usize, data: u8) {
-        assert!(addr < self.ram.len());
-        self.ram[addr] = data;
+    fn write_u8(&mut self, addr: u16, data: u8) {
+        assert!((addr as usize) < self.ram.len());
+        self.ram[addr as usize] = data;
     }
 }
