+++ src/nes/wram.rs
@@ -7,13 +7,13 @@ pub struct WorkRam {
 }
 
 impl SystemBus for WorkRam {
-    fn read_u8(&self, addr: usize) -> u8 {
-        assert!(addr < self.ram.len());
-        self.ram[addr]
+    fn read_u8(&self, addr: u16) -> u8 {
+        assert!((addr as usize) < self.ram.len());
+        self.ram[addr as usize]
     }
-    fn write_u8(&mut self, addr: usize, data: u8) {
-        assert!(addr < self.ram.len());
-        self.ram[addr] = data;
+    fn write_u8(&mut self, addr: u16, data: u8) {
+        assert!((addr as usize) < self.ram.len());
+        self.ram[addr as usize] = data;
     }
 }
